remove common test dependency
